app/requirements.txt
pillow==6.2.1
redis
eventlet
flask
flask_restful
Flask-PyMongo
flask-mongoengine
Flask-APScheduler
flask-cors
Werkzeug
Flask-SocketIO
wheel
influxdb-client
ipython
debugpy
requests
geopy
pydantic
flask_restful_swagger
flask_restful_swagger_3
fastapi
psycopg2-binary
sqlmodel
uvicorn
python-multipart

